requirements.txt
asgiref==3.2.7
Django==3.0.6
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
Flask==0.8
Jinja2==2.6
Werkzeug==0.8.3
certifi==0.0.8
chardet==1.0.1
distribute==0.6.24
gunicorn==0.14.2
requests==0.11.1
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pylint==2.5.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
virtualenv==20.0.18
whitenoise==5.0.1
wrapt==1.12.1
zipp==3.1.0